requirements.txt
asgiref==3.8.1
attrs==24.2.0
diploma-frontend @ file:///C:/Coding/Python/skillbox/python_django_diploma/diploma-frontend/dist/diploma-frontend-0.6.tar.gz
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
generic==1.1.3
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pillow==10.4.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
